+++ packages/drivers/odsp-driver/src/odspSnapshotParser.ts
@@ -66,7 +66,7 @@ export function convertOdspSnapshotToSnapsohtTreeAndBlobs(
     }
     const val: ISnapshotContents = {
         blobs: blobsWithBufferContent,
-        ops: odspSnapshot.ops ?? [],
+        ops: odspSnapshot.ops?.map((op) => op.op) ?? [],
         sequenceNumber: odspSnapshot.trees && (odspSnapshot.trees[0]).sequenceNumber,
         snapshotTree: buildHierarchy(odspSnapshot.trees[0]),
     };
